requirements.txt
alembic==1.9.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
